pyproject.toml
[project]
name = "dsst_etl"
dynamic = ["version"]
description = "An ETL package for data processing at DSST"
requires-python = ">=3.11"


dependencies = [
    "sqlalchemy",
    "psycopg2-binary",
    "python-dotenv",
    "pydantic",
    "alembic",
    "requests",
    "pandas",
    "metapub",
    "beautifulsoup4",
    "psycopg2-binary",
    "word2number",
    "pip",
    "pre-commit>=4.0.1",
    "pytest>=8.3.3",
    "boto3",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "black",
    "flake8",
    "mypy",
    "isort",
    "ipython",
    "types-requests",
    "pandas-stubs",
    "types-beautifulsoup4",
    "types-psycopg2",
]


[build-system]
requires = ["setuptools>=42", "wheel", "setuptools_scm[toml]>=3.4"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "dsst_etl/_version.py"

[tool.setuptools.packages.find]
where = ["."]

[project.scripts]
ingest = "dsst_etl.ingest:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = [ "py311"]

[tool.ruff]
line-length = 88
indent-width = 4
